Separate the struct CardViewCreator from the ViewController
It is renamed as CardViewFactory.
